Ignore system http_proxy when talking to ec2 metadata service
If http_proxy is set in the environment, requests will automatically
respect it - which is great if it's needed for talking to the aws API
endpoints, but causes great confusion when aurproxy tries to look up
the local system's instance ID in the ec2 metadata service.  This
change corrects the result to actually reflect the local machine, not
the instance ID of the http proxy.